% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sepMillier}
\alias{sepMillier}
\title{Separateur de millier}
\usage{
sepMillier(x)
}
\arguments{
\item{x}{a numeric value}
}
\value{
a string value of the input value with space thousand separator
}
\description{
Separateur de millier
}
\examples{
sepMillier(2340000)
# ouptut: 2 340 000
}
