% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_val.R
\name{square_val}
\alias{square_val}
\title{Square value}
\usage{
square_val(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
square value of the input
}
\description{
Square value
}
\examples{
square_val(2)
# output: 4
}
