% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{valueUnite}
\alias{valueUnite}
\title{Value Unite}
\usage{
valueUnite(val)
}
\arguments{
\item{val}{a numeric value}
}
\value{
string value of the input value, abbreviated with number ISU (K for Thousand, M for Million and MDS for Billion)
}
\description{
Value Unite
}
\examples{
valueUnite(2800)
output: 2.8 K
valueUnite(2370000)
output: 2.37 M
valueUnite(7890000000)
output: 7.89 MDS
}
